slackclient version update
