+++ environment.yml
@@ -3,6 +3,7 @@ channels:
   - defaults
 dependencies:
   - certifi=2019.9.11=py36_0
+  - pip=19.2.3=py36_0
   - python=3.6.9=h5500b2f_0
   - setuptools=27.2.0=py36_1
   - sqlite=3.29.0=he774522_0
@@ -10,14 +11,20 @@ dependencies:
   - vs2015_runtime=14.16.27012=hf0eaf9b_0
   - wheel=0.33.6=py36_0
   - pip:
+    - aiohttp==3.6.1
+    - async-timeout==3.0.1
+    - attrs==19.1.0
     - chardet==3.0.4
     - idna==2.8
-    - pip==19.2.3
+    - idna-ssl==1.1.0
+    - multidict==4.5.2
     - psycopg2==2.8.3
     - requests==2.22.0
     - six==1.12.0
-    - slackclient==1.3.2
+    - slackclient==2.2.0
+    - typing-extensions==3.7.4
     - urllib3==1.25.6
     - websocket-client==0.54.0
+    - yarl==1.3.0
 prefix: C:\Users\Artyom\Anaconda3\envs\digest
